requirements.txt
asgiref==3.8.1
cloudinary==1.39.1
Django==3.2.25
django-cloudinary-storage==0.3.0
djangorestframework==3.12.4
Pillow==8.2.0
pytz==2024.1
sqlparse==0.4.4
